task/TRAIN-1094 Fix: Fixed styling on actionbuttons and restructured folders
