+++ src/features/auth/components/Register/Register.tsx
@@ -1,4 +1,4 @@
-import AuthForm from '@/components/AuthForm/AuthForm';
+import AuthForm from '@/components/Forms/AuthForm/AuthForm';
 
 const Register = () => {
   return (
